Best universities/requirements_backup.txt
Brotli==1.0.9
click==7.1.2
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
dash-dependencies
dash
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
matplotlib
numpy==1.19.4
pandas==1.1.4
plotly==4.12.0
python-dateutil==2.8.1
pytz==2020.4
retrying==1.3.3
six==1.15.0
sklearn-preprocessing
sklearn-decomposition
urllib
Werkzeug==1.0.1
